data extraction from components
data extraction from components
