+++ pages/cycling.js
@@ -6,34 +6,7 @@ import Cover from '../components/cover';
 import BreadCrumb from '../components/breadBrumb';
 import { useTranslations } from 'use-intl';
 import pageService from '../services/pageService';
-
-const data = {
-  cover: {
-    src: '/bersac chambres dhotes.jpg',
-    textClass: 'home',
-  },
-  cta: {
-    minHeight: 150,
-    items: [
-      'accomodation',
-      'local',
-      'lakesAndRivers',
-      'watersports',
-      'trekking',
-      'moreActivities',
-    ]
-  },
-  cycleRoutes: [
-    '257955911',
-    '257963270',
-    '257964824',
-    '258012932'
-  ],
-  breadCrumb: [
-    { text: `Beaux Vue, Chambres d'Hotes`, link: '/' },
-    { text: 'Cycling' }
-  ]
-}
+import data from '../services/dataService';
 
 export async function getStaticProps({ locale }) {
   return {
