+++ pages/kayaking.js
@@ -5,28 +5,7 @@ import Cover from '../components/cover';
 import BreadCrumb from '../components/breadBrumb';
 import { useTranslations } from 'use-intl';
 import pageService from '../services/pageService';
-
-const data = {
-  cover: {
-    src: '/bersac chambres dhotes.jpg',
-    textClass: 'home',
-  },
-  cta: {
-    minHeight: 135,
-    items: [
-      'accomodation',
-      'local',
-      'watersports',
-      'lakesAndRivers',
-      'sup',
-      'moreActivities',
-    ]
-  },
-  breadCrumb: [
-    { text: `Beaux Vue, Chambres d'Hotes`, link: '/' },
-    { text: 'Kayaking' }
-  ]
-}
+import data from '../services/dataService';
 
 export async function getStaticProps({ locale }) {
   return {
