+++ pages/more-activities.js
@@ -5,34 +5,7 @@ import Cover from '../components/cover';
 import BreadCrumb from '../components/breadBrumb';
 import { useTranslations } from 'use-intl';
 import pageService from '../services/pageService';
-
-const data = {
-  cover: {
-    src: '/bersac chambres dhotes.jpg',
-    textClass: 'home',
-  },
-  cta: {
-    minHeight: 150,
-    items: [
-      'local',
-      'watersports',
-      'trekking',
-      'crafts',
-      'paragliding',
-      'xmass',
-      'cycling',
-      'fishing',
-      'kayaking',
-      'sup',
-      'windsurfing',
-      'wingSurfing'
-    ]
-  },
-  breadCrumb: [
-    { text: `Beaux Vue, Chambres d'Hotes`, link: '/' },
-    { text: 'More Activities' }
-  ]
-}
+import data from '../services/dataService';
 
 export async function getStaticProps({ locale }) {
   return {
