+++ pages/water-sports-rental.js
@@ -5,29 +5,7 @@ import Cover from '../components/cover';
 import BreadCrumb from '../components/breadBrumb';
 import { useTranslations } from 'use-intl';
 import pageService from '../services/pageService';
-
-const data = {
-  cover: {
-    src: '/bersac chambres dhotes.jpg',
-    textClass: 'home',
-  },
-  cta: {
-    minHeight: 135,
-    items: [
-      'accomodation',
-      'local',
-      'watersports',
-      'lakesAndRivers',
-      'kayaking',
-      'moreActivities',
-    ]
-  },
-  breadCrumb: [
-    { text: `Beaux Vue, Chambres d'Hotes`, link: '/' },
-    { text: 'Water Sports', link: '/watersports' },
-    { text: 'Water Sports Rental' },
-  ]
-}
+import data from '../services/dataService';
 
 export async function getStaticProps({ locale }) {
   return {
