+++ services/dataService.js
@@ -1,0 +1,61 @@
+import accomodation from './schemaFiles/accomodation';
+import cycling from './schemaFiles/cycling';
+import etangDeJonas from './schemaFiles/etangDeJonas';
+import etangDeSagnat from './schemaFiles/etangDeSagnat';
+import fishing from './schemaFiles/fishing';
+import gartempe from './schemaFiles/gartempe';
+import home from './schemaFiles/home';
+import kayaking from './schemaFiles/kayaking';
+import lacDeSaintPardoux from './schemaFiles/lacDeSaintPardoux';
+import lacDeVassivière from './schemaFiles/lacDeVassivière';
+import lacDuPontàLÂge from './schemaFiles/lacDuPontàLÂge';
+import laJonchèreSaintMaurice from './schemaFiles/laJonchèreSaintMaurice';
+import lakesAndRivers from './schemaFiles/lakesAndRivers';
+import leTaurionChatelusLeMarcheix from './schemaFiles/leTaurionChatelusLeMarcheix';
+import leTaurionPontDuDognon from './schemaFiles/leTaurionPontDuDognon';
+import local from './schemaFiles/local';
+import marsac from './schemaFiles/marsac';
+import moreActivities from './schemaFiles/moreActivities';
+import paragliding from './schemaFiles/paragliding';
+import saintSulpiceLauriereLake from './schemaFiles/saintSulpiceLauriereLake';
+import sup from './schemaFiles/sup';
+import trekking from './schemaFiles/trekking';
+import vienne from './schemaFiles/vienne';
+import waterSportsRental from './schemaFiles/waterSportsRental';
+import waterSports from './schemaFiles/waterSports';
+import windsurfing from './schemaFiles/windsurfing';
+import wingSurfing from './schemaFiles/wingSurfing';
+import xmass from './schemaFiles/xmass';
+
+const data = {
+  accomodation,
+  xmass,
+  cycling,
+  etangDeJonas,
+  etangDeSagnat,
+  fishing,
+  gartempe,
+  home,
+  kayaking,
+  laJonchèreSaintMaurice,
+  lacDeSaintPardoux,
+  lacDeVassivière,
+  lacDuPontàLÂge,
+  lakesAndRivers,
+  leTaurionChatelusLeMarcheix,
+  leTaurionPontDuDognon,
+  local,
+  marsac,
+  moreActivities,
+  paragliding,
+  saintSulpiceLauriereLake,
+  sup,
+  trekking,
+  vienne,
+  waterSportsRental,
+  waterSports,
+  windsurfing,
+  wingSurfing
+}
+
+export default data;
