+++ services/schemaFiles/cycling.js
@@ -1,0 +1,27 @@
+module.exports = {
+  cover: {
+    src: '/bersac chambres dhotes.jpg',
+    textClass: 'home',
+  },
+  cta: {
+    minHeight: 150,
+    items: [
+      'accomodation',
+      'local',
+      'lakesAndRivers',
+      'watersports',
+      'trekking',
+      'moreActivities',
+    ]
+  },
+  cycleRoutes: [
+    '257955911',
+    '257963270',
+    '257964824',
+    '258012932'
+  ],
+  breadCrumb: [
+    { text: `Beaux Vue, Chambres d'Hotes`, link: '/' },
+    { text: 'Cycling' }
+  ]
+};
