+++ services/schemaFiles/etangDeJonas.js
@@ -1,0 +1,39 @@
+module.exports = {
+  cover: {
+    src: '/lakes-and-rivers-france.jpg',
+    textClass: 'lakes-and-rivers',
+  },
+  cta: {
+    minHeight: 135,
+    items: [
+      'accomodation',
+      'fishing',
+      'watersports',
+      'cycling',
+      'trekking',
+      'moreActivities',
+    ]
+  },
+  lake: {
+    breadCrumb: [
+      { text: `Beaux Vue, Chambres d'Hotes`, link: '/' },
+      { text: 'Lakes and Rivers', link: '/lakes-and-rivers' },
+      { text: 'Etang De Jonas' },
+    ],
+    beach: false,
+    cycling: false,
+    fishing: true,
+    inflatables: false,
+    kayaking: false,
+    swimming: false,
+    trekking: false,
+    windsurfing: false,
+    wingSurfAndFoil: false,
+    gallery: [],
+    para:
+      <>
+        <p>
+        </p>
+      </>
+  }
+};
