+++ services/schemaFiles/fishing.js
@@ -1,0 +1,21 @@
+module.exports = {
+  cover: {
+    src: '/bersac chambres dhotes.jpg',
+    textClass: 'home',
+  },
+  cta: {
+    minHeight: 150,
+    items: [
+      'accomodation',
+      'local',
+      'lakesAndRivers',
+      'watersports',
+      'trekking',
+      'moreActivities',
+    ]
+  },
+  breadCrumb: [
+    { text: `Beaux Vue, Chambres d'Hotes`, link: '/' },
+    { text: 'Fishing' }
+  ]
+};
