+++ services/schemaFiles/home.js
@@ -1,0 +1,17 @@
+module.exports = {
+  cover: {
+    src: '/bersac chambres dhotes.jpg',
+    textClass: 'home',
+  },
+  cta: {
+    minHeight: 150,
+    items: [
+      'accomodation',
+      'local',
+      'watersports',
+      'fishing',
+      'cycling',
+      'moreActivities',
+    ]
+  }
+};
