+++ services/schemaFiles/kayaking.js
@@ -1,0 +1,21 @@
+module.exports = {
+  cover: {
+    src: '/bersac chambres dhotes.jpg',
+    textClass: 'home',
+  },
+  cta: {
+    minHeight: 135,
+    items: [
+      'accomodation',
+      'local',
+      'watersports',
+      'lakesAndRivers',
+      'sup',
+      'moreActivities',
+    ]
+  },
+  breadCrumb: [
+    { text: `Beaux Vue, Chambres d'Hotes`, link: '/' },
+    { text: 'Kayaking' }
+  ]
+};
