+++ services/schemaFiles/xmass.js
@@ -1,0 +1,21 @@
+module.exports = {
+  cover: {
+    src: '/bersac chambres dhotes.jpg',
+    textClass: 'home',
+  },
+  cta: {
+    minHeight: 150,
+    items: [
+      'accomodation',
+      'local',
+      'lakesAndRivers',
+      'crafts',
+      'trekking',
+      'moreActivities',
+    ]
+  },
+  breadCrumb: [
+    { text: `Beaux Vue, Chambres d'Hotes`, link: '/' },
+    { text: 'Paragliding' }
+  ]
+}
